add typescript support
